moved test custom DataObjects into a data folder in utilities project
