+++ src/System.Windows.Forms/tests/TestUtilities/Data/SerializableTestData.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+namespace System.Windows.Forms.TestUtilities;
+
+[Serializable]
+public class SerializableTestData
+{
+    public string Text { get; } = "a";
+}
